[DEV-1882] Integrated new pricing service
