+++ src/subdomains/core/buy-crypto/process/services/buy-crypto-preparation.service.ts
@@ -8,7 +8,7 @@ import { BankDataService } from 'src/subdomains/generic/user/models/bank-data/ba
 import { FeeService } from 'src/subdomains/supporting/payment/services/fee.service';
 import { TransactionHelper } from 'src/subdomains/supporting/payment/services/transaction-helper';
 import { Price } from 'src/subdomains/supporting/pricing/domain/entities/price';
-import { PriceProviderService } from 'src/subdomains/supporting/pricing/services/price-provider.service';
+import { PricingService } from 'src/subdomains/supporting/pricing/services/pricing.service';
 import { Between, In, IsNull, Not } from 'typeorm';
 import { BuyCryptoFee } from '../entities/buy-crypto-fees.entity';
 import { BuyCrypto } from '../entities/buy-crypto.entity';
@@ -24,7 +24,7 @@ export class BuyCryptoPreparationService {
   constructor(
     private readonly buyCryptoRepo: BuyCryptoRepository,
     private readonly transactionHelper: TransactionHelper,
-    private readonly priceProviderService: PriceProviderService,
+    private readonly pricingService: PricingService,
     private readonly fiatService: FiatService,
     private readonly bankDataService: BankDataService,
     private readonly buyCryptoWebhookService: BuyCryptoWebhookService,
@@ -70,8 +70,8 @@ export class BuyCryptoPreparationService {
           entity.paymentMethod,
         );
 
-        const inputAssetEurPrice = await this.priceProviderService.getPrice(inputReferenceCurrency, fiatEur);
-        const inputAssetChfPrice = await this.priceProviderService.getPrice(inputReferenceCurrency, fiatChf);
+        const inputAssetEurPrice = await this.pricingService.getPrice(inputReferenceCurrency, fiatEur, false);
+        const inputAssetChfPrice = await this.pricingService.getPrice(inputReferenceCurrency, fiatChf, false);
 
         const bankData = await this.bankDataService.getActiveBankDataWithIban(entity.bankTx.iban);
 
@@ -151,8 +151,8 @@ export class BuyCryptoPreparationService {
           entity.paymentMethod,
         );
 
-        const referenceEurPrice = await this.priceProviderService.getPrice(inputReferenceCurrency, fiatEur);
-        const referenceChfPrice = await this.priceProviderService.getPrice(inputReferenceCurrency, fiatChf);
+        const referenceEurPrice = await this.pricingService.getPrice(inputReferenceCurrency, fiatEur, false);
+        const referenceChfPrice = await this.pricingService.getPrice(inputReferenceCurrency, fiatChf, false);
 
         for (const feeId of fee.fees) {
           await this.feeService.increaseTxUsages(feeId, entity.user.userData);
