+++ src/subdomains/core/buy-crypto/process/services/buy-crypto-pricing.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@nestjs/common';
 import { Asset } from 'src/shared/models/asset/asset.entity';
 import { FeeResult } from 'src/subdomains/supporting/payout/interfaces';
-import { PriceProviderService } from 'src/subdomains/supporting/pricing/services/price-provider.service';
+import { PricingService } from 'src/subdomains/supporting/pricing/services/pricing.service';
 
 @Injectable()
 export class BuyCryptoPricingService {
-  constructor(private readonly priceProvider: PriceProviderService) {}
+  constructor(private readonly pricingService: PricingService) {}
 
   //*** PUBLIC API ***//
 
@@ -16,7 +16,7 @@ export class BuyCryptoPricingService {
   //*** HELPER METHODS ***//
 
   private async convertToTargetAsset(sourceAsset: Asset, sourceAmount: number, targetAsset: Asset): Promise<number> {
-    const price = await this.priceProvider.getPrice(sourceAsset, targetAsset);
+    const price = await this.pricingService.getPrice(sourceAsset, targetAsset, false);
 
     return price.convert(sourceAmount, 8);
   }
