--- src/subdomains/supporting/pricing/domain/enums/index.ts
@@ -1,65 +1,0 @@
-export enum PriceRequestContext {
-  BUY_CRYPTO = 'BuyCrypto',
-  REF_REWARD = 'RefReward',
-  PAY_IN = 'PayIn',
-}
-
-export enum Fiat {
-  EUR = 'EUR',
-  CHF = 'CHF',
-  USD = 'USD',
-  GBP = 'GBP',
-}
-
-export enum UsdStableCoin {
-  USDC = 'USDC',
-  USDT = 'USDT',
-  BUSD = 'BUSD',
-}
-
-export enum ChfStableCoin {
-  XCHF = 'XCHF',
-  ZCHF = 'ZCHF',
-}
-
-export enum Bitcoin {
-  BTC = 'BTC',
-  BTCB = 'BTCB',
-  WBTC = 'WBTC',
-}
-
-export enum AltCoin {
-  ETH = 'ETH',
-  WETH = 'WETH',
-  BNB = 'BNB',
-  MATIC = 'MATIC',
-  DAI = 'DAI',
-  UNI = 'UNI',
-  LINK = 'LINK',
-  ARB = 'ARB',
-  QNT = 'QNT',
-  APE = 'APE',
-  SAND = 'SAND',
-  MANA = 'MANA',
-  AAVE = 'AAVE',
-  GRT = 'GRT',
-  AXS = 'AXS',
-  RPL = 'RPL',
-  CRV = 'CRV',
-  BAT = 'BAT',
-  CHZ = 'CHZ',
-  COMP = 'COMP',
-  ENJ = 'ENJ',
-  MKR = 'MKR',
-  SNX = 'SNX',
-  TUSD = 'TUSD',
-  GMX = 'GMX',
-  INCH = '1INCH',
-  SUSHI = 'SUSHI',
-  XMR = 'XMR',
-}
-
-export enum SpecialCoin {
-  FOX = 'FOX',
-  BGB = 'BGB',
-}
