--- src/subdomains/supporting/pricing/domain/exceptions/path-misconfigured.exception.ts
@@ -1,5 +1,0 @@
-export class PathMisconfiguredException extends Error {
-  constructor(message: string) {
-    super(`PricePath init specification is not satisfied: ${message}`);
-  }
-}
