--- src/subdomains/supporting/pricing/domain/exceptions/path-not-configured.exception.ts
@@ -1,5 +1,0 @@
-export class PathNotConfiguredException extends Error {
-  constructor(from: string, to: string) {
-    super(`Pricing path is not configured for pair from: ${from} to: ${to} `);
-  }
-}
