--- src/subdomains/supporting/pricing/domain/exceptions/price-mismatch.exception.ts
@@ -1,5 +1,0 @@
-export class PriceMismatchException extends Error {
-  constructor(message: string) {
-    super(message);
-  }
-}
