--- src/subdomains/supporting/pricing/domain/exceptions/step-misconfigured.exception.ts
@@ -1,5 +1,0 @@
-export class StepMisconfiguredException extends Error {
-  constructor(message: string) {
-    super(`PriceStep init specification is not satisfied: ${message}`);
-  }
-}
