+++ src/subdomains/supporting/pricing/pricing.controller.ts
@@ -1,26 +1,14 @@
-import { Controller, Get, Put, Query, UseGuards } from '@nestjs/common';
+import { Controller, Put, UseGuards } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
 import { ApiBearerAuth, ApiExcludeEndpoint, ApiTags } from '@nestjs/swagger';
 import { RoleGuard } from 'src/shared/auth/role.guard';
 import { UserRole } from 'src/shared/auth/user-role.enum';
-import { PriceRequest, PriceResult } from './domain/interfaces';
 import { PricingService } from './services/pricing.service';
-import { PricingServiceNew } from './services/pricing.service.new';
 
 @ApiTags('pricing')
 @Controller('pricing')
 export class PricingController {
-  constructor(private readonly pricingService: PricingService, private readonly pricingServiceNew: PricingServiceNew) {}
-
-  @Get('price')
-  @ApiBearerAuth()
-  @ApiExcludeEndpoint()
-  @UseGuards(AuthGuard(), new RoleGuard(UserRole.ADMIN))
-  async getPrice(@Query() dto: PriceRequest): Promise<PriceResult> {
-    if (process.env.ENVIRONMENT === 'test') {
-      return this.pricingService.getPrice(dto);
-    }
-  }
+  constructor(private readonly pricingServiceNew: PricingService) {}
 
   @Put()
   @ApiBearerAuth()
