+++ src/subdomains/supporting/pricing/services/asset-prices.service.ts
@@ -6,7 +6,7 @@ import { FiatService } from 'src/shared/models/fiat/fiat.service';
 import { DfxLogger } from 'src/shared/services/dfx-logger';
 import { DisabledProcess, Process } from 'src/shared/services/process.service';
 import { Lock } from 'src/shared/utils/lock';
-import { PriceProviderService } from './price-provider.service';
+import { PricingService } from './pricing.service';
 
 @Injectable()
 export class AssetPricesService {
@@ -15,7 +15,7 @@ export class AssetPricesService {
   constructor(
     private readonly assetService: AssetService,
     private readonly fiatService: FiatService,
-    private readonly priceProvider: PriceProviderService,
+    private readonly pricingService: PricingService,
   ) {}
 
   // --- JOBS --- //
@@ -31,7 +31,7 @@ export class AssetPricesService {
 
     for (const asset of assetsToUpdate) {
       try {
-        const usdPrice = await this.priceProvider.getPrice(asset, usd);
+        const usdPrice = await this.pricingService.getPrice(asset, usd, false);
         await this.assetService.updatePrice(asset.id, usdPrice.convert(1));
       } catch (e) {
         this.logger.error(`Failed to update price of asset ${asset.uniqueName}:`, e);
