+++ src/subdomains/supporting/pricing/services/integration/currency.service.ts
@@ -1,24 +1,16 @@
 import { Injectable } from '@nestjs/common';
-import { PricingProvider } from '../../domain/interfaces';
-import { Price } from '../../domain/entities/price';
 import { HttpService } from 'src/shared/services/http.service';
-import { AsyncCache } from 'src/shared/utils/async-cache';
-import { Config } from 'src/config/config';
+import { Price } from '../../domain/entities/price';
+import { PricingProvider } from '../../domain/interfaces';
 
 @Injectable()
 export class CurrencyService implements PricingProvider {
   private readonly baseUrl = 'https://cdn.jsdelivr.net/gh/fawazahmed0/currency-api@1';
-  private readonly priceCache: AsyncCache<Price>;
 
-  readonly name: string;
-
-  constructor(private readonly http: HttpService) {
-    this.name = 'CurrencyService';
-    this.priceCache = new AsyncCache(Config.transaction.pricing.refreshRate * 60);
-  }
+  constructor(private readonly http: HttpService) {}
 
   async getPrice(from: string, to: string): Promise<Price> {
-    return this.priceCache.get(`${from}/${to}`, () => this.fetchPrice(from, to));
+    return this.fetchPrice(from, to);
   }
 
   private async fetchPrice(from: string, to: string): Promise<Price> {
